Move all tracking files to the tracking package
