+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/tracking/AddOrderTrackingProviderActionListener.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.orders
+package com.woocommerce.android.ui.orders.tracking
 
 interface AddOrderTrackingProviderActionListener {
     fun onTrackingProviderSelected(selectedCarrierName: String)
