+++ WooCommerce/src/main/res/layout/fragment_add_shipment_tracking.xml
@@ -5,7 +5,7 @@
     android:id="@+id/scroll_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.woocommerce.android.ui.orders.AddOrderShipmentTrackingFragment">
+    tools:context="com.woocommerce.android.ui.orders.tracking.AddOrderShipmentTrackingFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/snack_root"
