+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/orders/AddOrderTrackingProviderListPresenterTest.kt
@@ -10,6 +10,8 @@ import com.nhaarman.mockitokotlin2.whenever
 import com.woocommerce.android.network.ConnectionChangeReceiver.ConnectionChangeEvent
 import com.woocommerce.android.tools.NetworkStatus
 import com.woocommerce.android.tools.SelectedSite
+import com.woocommerce.android.ui.orders.tracking.AddOrderTrackingProviderListContract
+import com.woocommerce.android.ui.orders.tracking.AddOrderTrackingProviderListPresenter
 import org.junit.Before
 import org.junit.Test
 import org.wordpress.android.fluxc.Dispatcher
@@ -35,9 +37,11 @@ class AddOrderTrackingProviderListPresenterTest {
 
     @Before
     fun setup() {
-        presenter = spy(AddOrderTrackingProviderListPresenter(
+        presenter = spy(
+            AddOrderTrackingProviderListPresenter(
                 dispatcher, orderStore, wcStore, selectedSite, networkStatus
-        ))
+            )
+        )
         // Use a dummy selected site
         doReturn(SiteModel()).whenever(selectedSite).get()
         doReturn(true).whenever(networkStatus).isConnected()
